Navigation handled by code instead of implicitly by JavaFX and fxml.
The current navigation was handled in the fxml files using fxml include statements. I tried embedding 2 fxml files in one another, but I wanted to use the same controller for each fxml file, which JavaFx doesn't do if the controller is specified in the fxml. I removed the controllers from fxml, instead using coding the navigation myself.
Navigation was successful, however, it is going to be a huge overhead creating it myself. I'm going to navigation again to use the original method. Committing these changes in case I need them again.


 modified the navigation to use