Account Package Complete
Account Controller and Account Storage Handler successfully creates, updates, and deletes an account.

H2 documentation stated enums are stored as integers internally. AccountTypeEnum originally had a integer as a parameter to store the order of the enums in H2. However, the values is returned from H2 as a string. I've removed all methods referring to the integer).

Refactored CategoryController moving updating and deleting to their own methods instead of the onCommit methods.

Removed redundant try clauses in CategoryStorageHandler. The calling methods will deal with the exceptions.