Abstract method in EditableTextFieldTableCell allows extra validation
EditableTextFieldTableCell has an abstract method canUpdate checking if the new value in a modified cell can be updated.If it can, proceed as normal updating the model and saving the value to the database.

If it can't update, keep the cell active until a valid value is entered.

Label controller has implemented the new controller. Displays an alert with the error if the cell cannot be updated. Proceeds as noraml if it can be updated.